move MarkdownReader to components folder
