+++ src/pages/static/CookiesPolicy.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MarkdownReader from './MarkdownReader';
+import MarkdownReader from '../../components/markdownReader/MarkdownReader';
 import cookiesPolicy from './CookiesPolicy.md';
 
 function CookiesPolicy() {
