+++ src/pages/static/PrivacyPolicy.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MarkdownReader from './MarkdownReader';
+import MarkdownReader from '../../components/markdownReader/MarkdownReader';
 import privacyPolicy from './PrivacyPolicy.md';
 
 function PrivacyPolicy() {
