+++ src/pages/static/TermsAndConditions.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MarkdownReader from './MarkdownReader';
+import MarkdownReader from '../../components/markdownReader/MarkdownReader';
 import termsAndConditions from './TermsAndConditions.md';
 
 function TermsAndConditions() {
